requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-migrate>=4.1.0
flask-login>=0.6.3
flask-wtf>=1.3.0
email-validator>=2.3.0
python-dotenv>=1.0.0